requirements.txt
pythonnet>=3.0.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
